Adds logging framework.
This commit logs information from both the Sparrow Daemon
and the prototype backend to a special audit logging
file.

To support logging (and better request tracking in the
future), this commit also adds a requestId to the list
of parameters used when launching a task.  The requestId
is assigned by the Sparrow scheduler that launches the
job, and is unique across all schedulers.