Changing default AMI (instance ID) in deploy script.
For some reason ec2 is not correctly adding user public keys to the
authorized_keys list of deployed instances. This commit changes
the default AMI to a machine which already has Kay's public key so she
can run experiments. This is not a permenant solution - we still need to
figure out how to fix this.