Adding the notion of a task scheduler and a task launching service.
For the case when single nodes become overloaded, we want to add a way
to schedule tasks locally on one machine. This presents an alternative
to using containers, instead realying on Sparrow to queue tasks for each
user across all applications.

In the new architecture, the NodeMonitor submits a task to a TaskScheduler,
which is an abstract class providing a similar API to Java's BlockingQueue,
but with an unspecified (and therefore extensible) queuing policy. A
seperate service, called the TaskLauncherService, continuously waits for
new tasks to become runnable and launches them.