Added debugging to understand repsonse times for large simulations.
We were witnessing two pieces of unexplainable behavior; this
commit adds code to explain both pieces of behavior.  The first
piece of surprising behavior was that we were seeing surprisingly
long response times for random placement, even at low utilizations.
The second surprising thing was that the the response times for all
jobs in a given run of the simulation showed step-like behavior,
where most jobs finished at 100-millisecond intervals (for jobs
with 100-millisecond tasks).

To understand this behavior, this commit adds code to record the
wait time for tasks as a function of the result of the probe;
the graphs produced by this code clearly explain both behaviors.