Added the capability to model multi-core servers.
As part of doing this, removed part of the estimation for probes
that incorporated how long the current task had been running.
This estimation was a little unrealistic anyway, since in a real
deployment, we won't know the exactly length of each task. The
commit removes it because it's much more difficult to do this kind
of accounting for a multi-core machine.