Refactored scheduling code to use new scheduling mechanism.
Instead of sending probes that immediately return, and then scheduling tasks on the machines
with the lowest estimated load, the scheduler now sends reservations to the sample of node
monitors.  Node monitors don't respond until they have a free slot to run a task for the
reservation. The advantage of this new model is that we don't make any assumptions about
knowing task length.