Changed ProtoFrontend to allow constrained tasks, and fixed bugs.
The constrained task placer included many bugs, including issues where the wrong port
was used to describe a node monitor, issues with hostnames that were resolved
differently so didn't compare equal, and a bug in the allResponsesReceived() method.
This commit fixes those bugs. As of this commit,
the constrained task placer works with the ProtoFrontend in a 6-node deployment.