Changed start scripts to use concurrent mark/sweep GC for proto.
The proto frontend was pausing for long periods towards the end of
experiments running at higher load. This commit changes the
start_proto_frontend.sh script to use concurrent mark and sweep GC
for the proto frontend (just like we do for sparrow), and also adds
GC logging, in an attempt to fix the problem.