Fixed bug with round robin scheduler.
Previously, when the round robin task scheduler tried to launch a task for a
given user and all tasks in that job had already been launched, the user lost
that particular spot. This commit changes that behavior so that instead, if
all tasks in the particular job were launched, the scheduler tries to launch
another task for the same user.